+++ src/redteam/plugins/harmful/common.ts
@@ -1,18 +1,12 @@
 import { getEnvBool, getEnvString } from '../../../envars';
-import { PromptfooHarmfulCompletionProvider } from '../../../providers/promptfoo';
-import type { Assertion, AssertionType, PluginActionParams, TestCase } from '../../../types';
-import { sampleArray } from '../../../util/generation';
-import { sleep } from '../../../util/time';
-import type { UNALIGNED_PROVIDER_HARM_PLUGINS } from '../../constants';
+import type { Assertion, AssertionType, TestCase } from '../../../types';
 import {
   HARM_PLUGINS,
   LLAMA_GUARD_ENABLED_CATEGORIES,
   LLAMA_GUARD_REPLICATE_PROVIDER,
 } from '../../constants';
 
-export const PLUGIN_ID = 'promptfoo:redteam:harmful';
-
-export function getModerationAssertions(harmCategory: keyof typeof HARM_PLUGINS): Assertion[] {
+export function getHarmfulAssertions(harmCategory: keyof typeof HARM_PLUGINS): Assertion[] {
   const assertionType: AssertionType =
     harmCategory === 'harmful:privacy'
       ? 'promptfoo:redteam:harmful:privacy'
@@ -49,28 +43,6 @@ export function createTestCase(
       harmCategory: harmCategoryLabel,
     },
     metadata: { harmCategory: harmCategoryLabel },
-    assert: getModerationAssertions(harmCategory),
+    assert: getHarmfulAssertions(harmCategory),
   };
-}
-
-export async function getHarmfulTests(
-  { purpose, injectVar, n, delayMs = 0 }: PluginActionParams,
-  plugin: keyof typeof UNALIGNED_PROVIDER_HARM_PLUGINS,
-): Promise<TestCase[]> {
-  const unalignedProvider = new PromptfooHarmfulCompletionProvider({
-    purpose,
-    n,
-    harmCategory: plugin,
-  });
-  const tests: TestCase[] = [];
-  for (let i = 0; i < n; i++) {
-    const result = await unalignedProvider.callApi('');
-    if (result.output) {
-      tests.push(createTestCase(injectVar, result.output, plugin));
-    }
-    if (delayMs > 0) {
-      await sleep(delayMs);
-    }
-  }
-  return sampleArray<TestCase>(tests, n);
 }
