+++ src/redteam/plugins/index.ts
@@ -14,8 +14,8 @@ import { CrossSessionLeakPlugin } from './crossSessionLeak';
 import { DebugAccessPlugin } from './debugAccess';
 import { ExcessiveAgencyPlugin } from './excessiveAgency';
 import { HallucinationPlugin } from './hallucination';
-import { getHarmfulTests } from './harmful';
-import { AlignedHarmfulPlugin } from './harmful/alignedHarmful';
+import { AlignedHarmfulPlugin } from './harmful/aligned';
+import { getHarmfulTests } from './harmful/unaligned';
 import { ImitationPlugin } from './imitation';
 import { IntentPlugin } from './intent';
 import { OverreliancePlugin } from './overreliance';
